refactor code a bit to make usage of DriverAPI more explicit
because DriverAPI is not thread-safe, it’s better
to make its usage more explicit — so it’s now
passed as argument of all methods that need it.

also pass Engine as parameter in internal classes
that made little use of it (i.e. no point in
storing it if we can just pass it a sa parameter).

Change-Id: I9142959437ed3d8c8ecb7bcd030ea55d9c8ef352