Fix a bug when computing the warp-space alignment matrix
We where treating a direction as a vector when projecting the camera's
forward vector. We never saw this bug before because we always used
<0,0,0> as the origin of the light space.

Now that we can pick an arbitrary light-space, we use the camera's
origin. Other choices are possible. But the camera origin is a nice
reference point.

Also simplify some code.